refactor: made mouse control hook
